refactor: 인터셉터 클래스 commons 패키지에서 interceptor 로 변경
