+++ src/main/java/techcourse/myblog/config/WebConfig.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import techcourse.myblog.commons.AuthInterceptor;
-import techcourse.myblog.commons.LoginInterceptor;
+import techcourse.myblog.interceptor.AuthInterceptor;
+import techcourse.myblog.interceptor.LoginInterceptor;
 
 
 @Configuration
