+++ src/main/java/techcourse/myblog/interceptor/LoginInterceptor.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.commons;
+package techcourse.myblog.interceptor;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
