+++ src/test/java/techcourse/myblog/articles/ArticleTest.java
@@ -23,9 +23,9 @@ public class ArticleTest {
         String title = "title";
         String coverUrl = "coverUrl";
 //
-//        article.setContents(contents);
-//        article.setTitle(title);
-//        article.setCoverUrl(coverUrl);
+        article.setContents(contents);
+        article.setTitle(title);
+        article.setCoverUrl(coverUrl);
 
         assertThat(article.getContents()).isEqualTo(contents);
         assertThat(article.getTitle()).isEqualTo(title);
