+++ src/test/java/techcourse/myblog/interceptor/LoginInterceptorTest.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.commons;
+package techcourse.myblog.interceptor;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
